hsi2s: Implement DMA buffer read using current DMA address
An additional shared memory is used for the driver to update
the current write DMA address to the user application. Upon
the return of poll(), the user application reads this address
and then it store the contents from the previous address until
this address in the output file.

With this approach, even if the user application has missed
few interrupts from the driver, it can still read until the
latest DMA address as updated by the driver in the shared memory.

Change-Id: Id4de1bc53a5e42871522ef57e3ffb5c8eeedccbb